Chap2 - Pr3 & Pr5 (#2)
* Cracking the Coding Interview
Chapter 2, Problem 3
Implement an algorithm to delete a node from the middle of a linked list

* Cracking the Coding Interview
Chapter 2, Question 5
Sum of Lists
Question Text: You have two numbers represented by a linked list, where each node contains a single digit. These digits are stored in reverse order, such that the 1's digit is at the head of the list. Write a function that adds the two numbers and returns the sum as a linked list.

* Solved Chap 2 question 5, modified linked list utility classes & corresponding method names in ch2q3 solution

* Revised static method and added 1 additional test case to PSVM